Readmes, Moved files, Added text
