+++ README.md
@@ -1,2 +1,4 @@
-# sb_dictionary
-sb_dictionary
+# Schwäbisch Dictionärle
+* API and Frontend have their own ReadMes
+* [Workboard on Trello](hhttps://trello.com/b/t90SDmBW/schw%C3%A4bisch-diction%C3%A4rle)
+
