--- ionic/src/app/explore-container/explore-container.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-explore-container',
-  templateUrl: './explore-container.component.html',
-  styleUrls: ['./explore-container.component.scss'],
-})
-export class ExploreContainerComponent implements OnInit {
-  @Input() name: string;
-
-  constructor() { }
-
-  ngOnInit() {}
-
-}
