+++ ionic/src/app/pages/tabs/tab1/tab1.page.html
@@ -1,7 +1,7 @@
 <ion-header [translucent]="true">
   <ion-toolbar>
     <ion-title>
-    Land Page
+      Mahlzeit
     </ion-title>
   </ion-toolbar>
 </ion-header>
@@ -11,17 +11,18 @@
 
     <ion-slide>
        <img src="https://cdn.iconscout.com/icon/premium/png-256-thumb/q-and-a-1848732-1568044.png">
-      <h2>Welcomen zu <br>  <b>Wörter auf schwäbisch</b></h2>
+      <h2><b>Schwäbisch Dictionärle</b> für Neigschmeckte</h2>
       <p>
-        Lorem ipsum dolor sit amet, consetetur sadipscing elitr, accusam et
+          Heit gats um des Wörtle:<br>
+          <b>[TODO]</b>
       </p>
     </ion-slide>
     <ion-slide>
       <img src="https://cdn.iconscout.com/icon/premium/png-256-thumb/q-and-a-1848732-1568044.png"/>
-      <h2>What is Ionic?</h2>
+      <h2>Wasch'n des?</h2>
       <p>
-        Lorem ipsum dolor sit amet, consetetur sadipscing elitr, accusam et justo duo dolores et ea rebum.
-        Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
+        Wie ihr älle scho wisschd: Gottes schönste Gabe, ist und bleibt der Schwabe.
+        <br><br> So lernsch Schwäbisch schwätza!
       </p>
     </ion-slide>
   </ion-slides>
