+++ ionic/src/app/pages/tabs/tab1/tab1.page.scss
@@ -5,11 +5,8 @@
 .swiper-slide {
   display: block;
 }
-ion-slide > h2 {
-  margin-top: 2.8rem;
-}
+
 ion-slide > img {
   max-height: 40%;
   max-width: 60%;
-  margin: 36px 0;
 }
