+++ ionic/src/app/pages/tabs/tab1/tab1.page.ts
