+++ ionic/src/app/pages/tabs/tab2/detail/detail-routing.module.ts
