+++ ionic/src/app/pages/tabs/tab2/detail/detail.page.html
