+++ ionic/src/app/pages/tabs/tab2/sentence-list/sentence-list.component.html
