+++ ionic/src/app/pages/tabs/tab2/tab2-routing.module.ts
@@ -9,7 +9,7 @@ const routes: Routes = [
   },
   {
     path: ':id',
-    loadChildren: () => import('./detail/detail.module').then( m => m.DetailPageModule)
+    loadChildren: () => import('./detail/detail.module').then(m => m.DetailPageModule)
   }
 ];
