+++ ionic/src/app/pages/tabs/tab2/tab2.module.ts
@@ -4,7 +4,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Tab2Page } from './tab2.page';
-import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
 
 import { Tab2PageRoutingModule } from './tab2-routing.module';
 import {Ng2SearchPipeModule} from "ng2-search-filter";
@@ -14,7 +13,6 @@ import {Ng2SearchPipeModule} from "ng2-search-filter";
     IonicModule,
     CommonModule,
     FormsModule,
-    ExploreContainerComponentModule,
     Tab2PageRoutingModule,
     Ng2SearchPipeModule,
   ],
