+++ ionic/src/app/pages/tabs/tab3/tab3.page.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
-import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
+import { ExploreContainerComponentModule } from '../tab1/explore-container/explore-container.module';
 
 import { Tab3Page } from './tab3.page';
