+++ ionic/src/app/services/word.service.ts
@@ -1,10 +1,9 @@
 import { Injectable } from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {environment} from '../../environments/environment';
-import {IWord} from "../models/word";
+import {IWord} from "../models/IWord";
 import { Observable, of } from 'rxjs';
 import { catchError, tap } from 'rxjs/operators';
-
 @Injectable({
   providedIn: 'root'
 })
@@ -20,8 +19,8 @@ export class WordService {
   }
 
   public getWord(id: string): Observable<IWord[]> {
-    console.log(`${environment.apiLocation}api/words/`+id);
-    return this.http.get<IWord[]>(`${environment.apiLocation}api/words/`+id);
+    console.log(`${environment.apiLocation}api/words/${id}`);
+    return this.http.get<IWord[]>(`${environment.apiLocation}api/words/${id}`);
   }
 
 }
